+++ MimeKit/Tnef/TnefPart.cs
@@ -245,12 +245,7 @@ namespace MimeKit.Tnef {
 								var tnef = new TnefPart ();
 
 								foreach (var param in attachment.ContentType.Parameters)
-									if (!tnef.ContentType.Parameters.Contains (param.Name)) {
-										// It is not clear that duplicate content-type paramater names are illegal,
-										// but we should not break down when confronted with them.
-										// https://tools.ietf.org/html/rfc2231
-										tnef.ContentType.Parameters.Add (param.Name, param.Value);
-									}
+									tnef.ContentType.Parameters[param.Name] = param.Value;
 
 								if (attachment.ContentDisposition != null)
 									tnef.ContentDisposition = attachment.ContentDisposition;